+++ app/indexers/spot/collection_indexer.rb
@@ -2,5 +2,15 @@
 module Spot
   class CollectionIndexer < Hyrax::CollectionIndexer
     include IndexesLanguageAndLabel
+
+    self.thumbnail_path_service = ::Spot::CollectionThumbnailPathService
+
+    def generate_solr_document
+      super.tap do |doc|
+        if (slug_id = object.identifier.find { |id| id.start_with? 'slug:' })
+          doc['collection_slug_ssi'] = Spot::Identifier.from_string(slug_id).value
+        end
+      end
+    end
   end
 end