+++ app/presenters/spot/collection_presenter.rb
@@ -7,6 +7,20 @@ module Spot
 
     delegate :abstract, :related_resource, to: :solr_document
 
+    # Presenter fields displayed on the #show sidebar (on the right).
+    # Modify this to change what's displayed + the order.
+    #
+    # @return [Array<Symbol>]
+    def self.terms
+      [
+        :total_items,
+        :related_resource,
+        :location,
+        :sponsor,
+        :modified_date
+      ]
+    end
+
     # @return [true,false]
     def collection_featurable?
       user_can_feature_collections? && solr_document.public?
@@ -28,9 +42,12 @@ module Spot
       @featured
     end
 
-    # @todo this is a no-op at the moment. we'll need to update the metadata to provide
-    #       a URL that provides more information about the collection
-    def learn_more_url; end
+    # Is the document's visibility public?
+    #
+    # @return [true, false]
+    def public?
+      solr_document.visibility == ::Hydra::AccessControls::AccessRight::VISIBILITY_TEXT_VALUE_PUBLIC
+    end
 
     # @return [true,false]
     def user_can_feature_collections?